Split points into points and base points
